% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologize.R, R/homologize_core.R
\name{homologize_core}
\alias{homologize_core}
\alias{homologize}
\title{Homologize Core Function}
\usage{
homologize(
  project_name = NULL,
  seurat_dir = NULL,
  organism_in = NULL,
  organism_out = NULL,
  check_umap = NULL,
  output_dir = NULL,
  sample_list = NULL
)
}
\arguments{
\item{project_name}{String with project or sample name.}

\item{seurat_dir}{Path to the Seurat object used as input}

\item{organism_in}{Organism of the input Seurat object (Hsap, Mmul, Mmus)}

\item{organism_out}{Organism for which gene homologs must be found}

\item{check_umap}{Whether or not marker umaps shoudl be generated. Default is TRUE.}

\item{output_dir}{Directory where homologize_core output should be generated.}

\item{sample_list}{Input multiple samples in list. Default is NULL}
}
\value{
Seurat object

Seurat object
}
\description{
Helper function responsible for converting genes in a Seurat object
to homologous genes from another organism.

Helper function responsible for converting genes in a Seurat object
to homologous genes from another organism.
}
\examples{
\dontrun{
  seurat_obj <- homologize(
    project_name = "example_project",
    seurat_dir   = "path/to/seurat_object.rds",
    organism_in  = "human",
    organism_out = "mouse",
    check_umap   = TRUE,
    output_dir   = "path/to/output"
  )
}
\dontrun{
  seurat_obj <- homologize_core(
    project_name  = "example_project",
    seurat_dir    = "path/to/seurat_object.rds",
    organism_in   = "human",
    organism_out  = "mouse",
    check_umap    = TRUE,
    output_dir    = "path/to/output"
  )
}
}
