% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_samples.R
\name{integrate_samples}
\alias{integrate_samples}
\title{Integrate Samples of Same Species Origin}
\arguments{
\item{project_name}{String with project name.}

\item{organism}{Which organism do all inputs belong to, either Hsap, Mmul, or Mmus.}

\item{sample_list}{Required list of directories for input Seurat objects.}

\item{output_dir}{Directory where integrate_samples output should be generated.}
}
\value{
Single, integrated Seurat object
}
\description{
Integrate Samples of Same Species Origin
}
\examples{
\dontrun{

 # Define input list, here all are defined in terms of mouse genes
  samples <- list(
    "/home/user/postalignemnt/Homologize/R_objects/human_1_mouse.rds",
    "/home/user/postalignemnt/Homologize/R_objects/human_2_mouse.rds",
    "/home/user/postalignemnt/Homologize/R_objects/human_3_mouse.rds",
    "/home/user/postalignemnt/Homologize/R_objects/human_4_mouse.rds"
    )

  # Run the integrate samples function
  output <- integrate_samples(
    project_name = "mouse_samples",
    organism = "mouse",
    sample_list = samples,
    output_dir = "/home/user/postalignment/integrate_samples/"
  )

}
}
