% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_species.R
\name{integrate_species}
\alias{integrate_species}
\title{Integrate Samples of Different Species Origin}
\arguments{
\item{project_name}{String with project name.}

\item{organism}{Organism that all samples gene annotations conform to.}

\item{sample_list}{Required list of directories for input Seurat objects.}

\item{method}{Integration method to use, either CCA or harmony, default is harmony.}

\item{output_dir}{Directory where integrate_species output should be generated.}
}
\value{
Single, integrated Seurat object
}
\description{
Integrate Samples of Different Species Origin
}
\examples{
\dontrun{
 # Define input list, here all are defined in terms of mouse genes
  samples <- list(
    "/home/user/postalignemnt/Homologize/R_objects/human_1_mouse.rds",
    "/home/user/postalignemnt/Homologize/R_objects/human_2_mouse.rds",
    "/home/user/postalignemnt/Homologize/R_objects/mouse_1_mouse.rds",
    "/home/user/postalignemnt/Homologize/R_objects/mouse_2_mouse.rds",
    "/home/user/postalignemnt/Homologize/R_objects/macaque_1_mouse.rds",
    "/home/user/postalignemnt/Homologize/R_objects/macaque_2_mouse.rds")

  # Run the integrate species function
  output <- integrate_species(
    project_name = "mouse_genes_harmony",
    organism = "mouse",
    sample_list = samples,
    method = "harmony",
    output_dir = "/home/user/postalignment/integrate_species/"
  )

}
}
