% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess Full Function}
\arguments{
\item{project_name}{String with project or sample name.}

\item{organism}{String with one of the following Mmul, Hsap, Mmus.}

\item{ptprc_threshold}{Percentage CD45 over which cells will be retained. E.g: If 0, all cells must express CD45}

\item{filtered_path}{Directory of filtered alignment output, matrix.mtx.gz, features.tsv.gz, barcodes.tsv.gz.}

\item{raw_path}{Directory of raw alignment output, matrix.mtx.gz, features.tsv.gz, barcodes.tsv.gz.}

\item{velocyto_path}{Directory of Velocyto filtered alignment output, spliced.mtx.gz, unspliced.mtx.gz, ambiguous.mtx.gz, features.tsv.gz, barcodes.tsv.gz.}

\item{filter_damaged}{Should output contain no damaged cells. Default is TRUE.}

\item{output_dir}{Directory where preprocess_core output should be generated.}

\item{sample_list}{Input multiple samples in list. Default is NULL}
}
\value{
Seurat object
}
\description{
Function responsible for heavy lifting preprocessing work.
This includes ambient RNA correction with SoupX, removing red blood cells,
isolating immune cells, filtering damaged cells detected by DropletQC and
limiric, and filtering doublets detected by DoubletFinder. Also provided
is preliminary check of sample quality by immune cell marker visualisation (UMAP).
}
\examples{
\dontrun{
  # Function can run multiple samples or a single sample

  # Run a single sample
  SRR1234567 <- preprocess_core(project_name = "SRR1234567",
                                filtered_path = "/home/username/alignment/SRR1234567/filtered",
                                raw_path = "/home/username/alignment/SRR1234567/raw",
                                velocyto_path = "/home/username/alignment/SRR1234567/velocyto",
                                output_dir = "/home/username/postalignemnt/SRR1234567/preprocess")

 # Run multiple samples
 samples <- list(
 list(project_name = "SRR1234567",
     filtered_path = "/home/username/alignment/SRR1234567/filtered",
     raw_path = "/home/username/alignment/SRR1234567/raw",
     velocyto_path = "/home/username/alignment/SRR1234567/velocyto",
     output_dir = "/home/username/postalignemnt/SRR1234567/preprocess"),

 list(project_name = "SRR1234568",
     filtered_path = "/home/username/alignment/SRR1234568/filtered",
     raw_path = "/home/username/alignment/SRR1234568/raw",
     velocyto_path = "/home/username/alignment/SRR1234568/velocyto",
     output_dir = "/home/username/postalignemnt/SRR1234568/preprocess"))

GSE1234567 <- preprocess(sample_list = samples)

}
}
